Update the test on user( which is the formerly token)
